PRO-3048: Renamed package to avoid conflicts
