+++ src/contractTest/java/uk/gov/hmcts/probate/contract/model/ClientAuthorizationCodeResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.probate.model;
+package uk.gov.hmcts.probate.contract.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
