+++ src/contractTest/resources/application.properties
@@ -12,5 +12,3 @@ idam.secret=${IDAM_SECRET:abc}
 s2s.auth.totp.secret=${S2S_AUTH_TOTP_SECRET:abc}
 idam.username=${IDAM_USERNAME:testusername}
 idam.userpassword=${IDAM_USERPASSWORD:testpassword}
-idam.s2s-auth.url=http://localhost:4501
-idam.password=password
